cra-planner-agent/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.12.0
requests>=2.31.0
duckduckgo-search>=4.1.0
python-dotenv>=1.0.0